backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/main_app"]

[project]
name = "fastapi-supabase-backend"
version = "0.1.0"
description = "FastAPI backend with Supabase integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "supabase>=2.4.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
fastapi-start = "main_app.main:main"

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true